provider crud completed

+++ app/Http/Controllers/Api/CategoryController.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Api;
 
+use App\Filters\Category\CategoryFilter;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
 use App\Models\CategoryModel;
@@ -17,32 +18,66 @@ class CategoryController extends Controller
      */
     public function index(Request $request): JsonResponse
     {
-        $categories = CategoryModel::query()->paginate($request->get('per_page', 10));
+        $filters[] = CategoryFilter::getRequest($request);
+
+        $categories = CategoryModel::applyEloquentFilters($filters)
+//            ->with('categories')
+            ->select('id', 'name', 'category_id')
+            ->paginate($request->get('per_page', 10));
         return $this->paginateRes($categories);
     }
 
+    /**
+     * @param CategoryRequest $request
+     * @return JsonResponse
+     */
     public function store(CategoryRequest $request): JsonResponse
     {
         CategoryModel::query()->create($request->validated());
         return $this->successRes();
     }
 
+    /**
+     * @param int $id
+     * @return JsonResponse
+     */
     public function getOne(int $id): JsonResponse
     {
-        return $this->successRes(CategoryModel::query()->findOrFail($id)->toArray());
+        $category = CategoryModel::query()->find($id);
+        if (!$category) {
+            return $this->errorRes('Category not found', 404, 404);
+        }
+        return $this->successRes($category->toArray());
     }
+
+    /**
+     * @param CategoryRequest $request
+     * @param int $id
+     * @return JsonResponse
+     */
 
     public function update(CategoryRequest $request, int $id): JsonResponse
     {
-        $category = CategoryModel::query()->findOrFail($id);
+        $category = CategoryModel::query()->find($id);
+        if (!$category) {
+            return $this->errorRes('Category not found', 404, 404);
+        }
         $category->fill($request->validated());
         $category->save();
         return $this->successRes();
     }
 
+    /**
+     * @param int $id
+     * @return JsonResponse
+     */
+
     public function destroy(int $id): JsonResponse
     {
-        $category = CategoryModel::query()->findOrFail($id);
+        $category = CategoryModel::query()->where('id', '=', $id)->first();
+        if (!$category) {
+            return $this->errorRes('Category not found', 404, 404);
+        }
         $category->delete();
         return $this->successRes();
     }
